Task #3 - Add persistent storage

+++ ConsoleBookingSystem/HostService.cs
@@ -1,7 +1,10 @@
+using System.Data;
+
 namespace ConsoleBookingSystem;
 
 public class HostService : IHostService
 {
+    private bool _anyUnsavedChanges = false;
     private IHostRepository _hostRepository;
 
     public HostService(IHostRepository hostRepository)
@@ -36,37 +39,56 @@ public class HostService : IHostService
     public void AddHost(string hostName, string address)
     {
         if (HostExistsByNameAndAddress(hostName, address))
-        {
-            throw new Exception("Host with specified name and address already exists.");
-        }
+            throw new ArgumentException("Host with specified name and address already exists.");
         
         _hostRepository.CreateHost(new CreateHostData { Name = hostName, Address = address });
+        
+        SetAnyUnsavedChanges(true);
     }
 
     public void RemoveHostById(int hostId)
     {
+        if (hostId <= 0)
+            throw new ArgumentOutOfRangeException(nameof(hostId), "Host ID cannot be zero or negative.");
+        
         if (_hostRepository.FindHostsCount() == 0)
-        {
-            throw new Exception("No Hosts added yet.");
-        }
+            throw new HostNotFoundException("No Hosts found.");
             
         var host = _hostRepository.FindHostById(hostId);
         if (host == null)
-        {
-            throw new Exception($"Host with ID = {hostId} not found.");
-        }
+            throw new HostNotFoundException("Host with specified ID not found.", hostId);
         
         _hostRepository.RemoveHost(host);
+        
+        SetAnyUnsavedChanges(true);
     }
 
     public void EditHostById(int hostId, string updatedHostName, string updatedAddress)
     {
         if (!HostExistsById(hostId))
-            throw new Exception($"Host with ID = {hostId} not found.");
+            throw new HostNotFoundException("Host with specified ID not found.", hostId);
 
         if (HostExistsByNameAndAddress(updatedHostName, updatedAddress, hostId))
-            throw new Exception("Host with specified Name and Address already exists.");
+            throw new ArgumentException("Host with specified Name and Address already exists.");
         
         _hostRepository.UpdateHost(hostId, new UpdateHostData { Name = updatedHostName, Address = updatedAddress });
+        
+        SetAnyUnsavedChanges(true);
+    }
+
+    public void SaveChanges()
+    {
+        _hostRepository.SaveChanges();
+        SetAnyUnsavedChanges(false);
+    }
+    
+    private void SetAnyUnsavedChanges(bool anyUnsavedChanges)
+    {
+        _anyUnsavedChanges = anyUnsavedChanges;
+    }
+    
+    public bool AnyUnsavedChanges()
+    {
+        return _anyUnsavedChanges;
     }
 }
